+++ aoc/src/main/kotlin/com/webnologies/grkernisant/aoc/aoc2024/day19/TowelPattern.kt
@@ -3,7 +3,7 @@ package com.webnologies.grkernisant.aoc.aoc2024.day19
 data class TowelPattern(
     override val pattern: String,
     override val blocks: MutableSet<String> = mutableSetOf(),
-    override var nbCombo: Int = 1,
+    override var nbCombo: Long = 1L,
     override val sets: MutableSet<String> = mutableSetOf(),
     var isUnique: Boolean = false,
 ) : PatternInterface {
